refactor(ui): split file diff tauri services
